Small fix to unittest harness
TestSuites should produce a success file even if in a given configuration, none of the tests will ever be run. This allows us to check for success simply in the presubmit without needing to know if any of the sub-tests actually run in a given configuration or not.
Review URL: http://codereview.appspot.com/4532088