Create ExplicitOrderer.
This CL creates an orderer that applies the order information specified by a Reorderer::Order object. All of this will eventually disappear in a post-metadata world, but this acts as a bridge between the old code (reorder_lib) and the new code (transforms and orderers) until block-graph metadata is available.

BUG=
TEST=

Review URL: https://codereview.appspot.com/5645049