Manually trim source range on DOS header rewriting.
This allows us to successfully instrument release EXE files, which would break due to some pretty interesting code generation.
The issue is that the EXE-specific CRT has code like this:
  PIMAGE_DOS_HEADER pDOSHeader;
  PIMAGE_NT_HEADERS pPEHeader;

  pDOSHeader = &__ImageBase;
  if ( pDOSHeader == NULL || pDOSHeader->e_magic != IMAGE_DOS_SIGNATURE )
    return 0;

  pPEHeader = (PIMAGE_NT_HEADERS)((char *)pDOSHeader +
               pDOSHeader->e_lfanew);

  if ( pPEHeader->Signature != IMAGE_NT_SIGNATURE )
    return 0;

  if (pPEHeader->OptionalHeader.Magic != IMAGE_NT_OPTIONAL_HDR_MAGIC)
    return 0;

The code generator optimizes this by storing pDOSHeader->e_lfanew in a register (EAX), and then performs the pPEHeader references with __ImageBase+OffsetOf(IMAGE_NT_HEADER, <field>) as a base pointer, e.g.

  210 01476cfe 81b80000310150450000 cmp dword ptr shell!__ImageBase (01310000)[eax],4550h

and

  210 01476d0f 66398818003101  cmp     word ptr shell!__ImageBase+0x18 (01310018)[eax],cx

This effectively means that we can have code containing literal pointers that refer to globals, but with offsets that don't relate to that global at all. Thankfully this will only occur in the case of data, as pointer arithmetic is not performed on function pointers.

R=chrisha@chromium.org,rogerm@chromium.org

Review URL: https://codereview.appspot.com/5651089