In a recent CL I introduced a bug by not detecting and properly handling double transitions of a buffer to 'pending write' state. This can occur when Close is called, and a call to ReturnBuffer is made prior to the session being destroyed. I would end up double incrementing the 'buffers_pending_write' counter, and later DCHECK on the unexpected value.
This CL makes bookkeeping more explicit by augmenting buffers with state information, and keeping track of all buffers in the session. It also makes it so that the session is the only place where buffer state is altered, rather than in Session and Service.

Review URL: https://codereview.appspot.com/5874050