Basic-Block Randomization.
This CL uses the BlockBuilder class to flesh out the ExplodeBasicBlocksTransform in order to perform basic-block randomization in the relinker application.

There are a few unresolved issues in the basic block decomposer that prevent randomization of chrome.dll. These will be addressed in a subsequent CL. In particlar, the way BasicBlockReferrer is modeled needs to be expanded to support Instructions and Successors are referrers and should stop copying intra-block referrers.

See: https://codereview.appspot.com/6354049/

BUG=
TEST=

Review URL: https://codereview.appspot.com/6243056