Yet another change to BlockGraphSerializer API.
Make the callbacks be called for every single block, rather than only those who need their data saved/loaded. This allows the callbacks to be used as a mechanism to save extra data along-side a block-graph as well (as in the case of block-graph + image-layout). This entailed a small API change to explicitly tell the callback whether or not it is also responsible for the block data.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6352058