Reading of the field list type info.
New parameter to the dump functions of the type info dump function allowing to specify the level of indentation.

All the type that can be encounter inside a field list are not supported yet but I've dumped some of them to make sure that the level of indentation mecanism work correctly.

I should maybe add this param to the symbol record dump functions.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6349059