Add trailing instruction bytes to assembly test functions.
This allows our decomposer to handle these blocks without error. Currently, we blow up because the '*_end' labels fall off the end of the block. (Alas, we could have done this with dos_stub.asm as well, if we want to revert that CL.)

BUG=

Review URL: https://codereview.appspot.com/6493083