Add 'spawn' command to call_trace_service.exe.
This causes call_trace_service to launch an instance of itself in the background using the 'start' command, and to wait for it to be ready (when it fires off an appropriate event), and then returns. This allows for reliable starting of the call-trace service from automated scripts.

BUG=

Review URL: https://codereview.appspot.com/6499109