Make DiaBrowser provide a 'pop' callback.
DiaBrowser provides a regex-like pattern-guided DFS search through the tree of symbols in a PDB file. This adds a 'pop' callback which indicates when a previously matched element is removed from the stack of symbols, allowing clients to maintain shadow stacks of metadata. To be used in the revamped decomposer.

BUG=

Review URL: https://codereview.appspot.com/6854120