Do not generate an error on partial memory read for a stack trace.
When generating a remote stack trace in the asan logger, it would
interpret a partial read as an error and generate a log message.
The StackWalk64 function seems to do the correct thing (presumably
it checks GetLastError()) whether this is returns as an error or
not, so the log message was spurious.

R= chrisha@chromium.org

Review URL: https://codereview.appspot.com/7241057