Fix a P0 bug in Decomposer.
In some extreme cases this was causing us to interpret jump tables as if they were valid instructions. This has not been a problem with Chrome or test_dll because we always generate full symbols. This CL includes the addition of a compiland that is added to test_dll with no private symbols.

BUG=

Review URL: https://codereview.appspot.com/8336043