Make the slow path faster !
We had a tendency to consider the slow path as a rare case but it turns out that this is not really the case... This CL inline the IsAccessible function into the asm routine. This bring a little bit more complexity but this make the runtime much faster ! I observe a gain between 0.24 and 1.34 x on a dromaeo testbench (our worst case in this benchmark is reduced from 4.76x to 3.42x). Maybe this should be ran on a CPU2006 benchmark before landing ?

BUG=
R=chrisha@chromium.org, siggi@chromium.org

Review URL: https://codereview.appspot.com/9027043