Reduce the memory overhead per block and add the TIDs to the header.
Instead of having a big redzone at the beginning and at the end of each blocks (at least 64 bytes in total) we now have a static overhead of 36 bytes split in two redzones, one in front of the block and the other one after it. The new layout of a memory access is as follow:
L L L L X X ... X X R R R R P
L: 4 byte of header redzone.
X: User allocation
R: 4 byte of trailer redzone.
P: Padding to be 8-byte aligned

The header and trailer are composed as follow:

- Magic number             24 bits
- State	                   8 bits(1)
- Allocation stack trace   32 bits
- Block size               32 bits
- Alloc TID                32 bits
                  Total:   128 bits

Trailer
- Timestamp                64 bits
- Free stack trace         32 bits
- Free TID                 32 bits
- Next free block          32 bits
                  Total:   160 bits

(1) We could use less bits here but as we want the redzone to be 8-byte aligned we've room for those extra bits.

We see a 2x improvement on the memory usage (overhead ratio = allocated bytes / requested bytes):
Trunk@r1510:
Browser:
Overhead ratio: 1.666697
Renderer:
Overhead ratio: 1.293409

New Block header version:
Browser:
Overhead ratio: 1.329258
Renderer:
Overhead ratio: 1.144337

Here's 3 runs of the dromaeo benchmark to measure the impact on the performances (this CL / trunk@r1510):
585.38 / 593.95
598.04 / 579.53
585.27 / 574.62
Avg of the 3 runs : 589.563 / 582.7

Those results are similar, so this doesn't seems to introduce a slowdown (we even see a 1-2% improvement but this can due to the imprecision of dromaeo...)

BUG=
R=siggi@chromium.org

Review URL: https://codereview.appspot.com/9400043