Improve ASan's symbolizer to support content_shell etc...
The problem was that this script was specific to chrome.dll, now that we have a split build and that clusterfuzz also use content_shell.exe we need can't simply do a "reload /fi chrome.dll" to force the symbols of chrome.dll to be loaded as chrome.dll might not be in the modules list.

Now I scan the loaded modules and the pdbs present in the |pdb_path| passed as an argument to the script and I load the appropriate symbols.

R=chrisha@chromium.org

BUG=

Review URL: https://codereview.appspot.com/65780045