Prevent agents from using SSE/SSE2/AVX instructions.
As of VS2013 the CRT makes use of SSE2 by default, and code is generated with SSE2 instructions. This was causing our instrumentation to stomp on SSE2 registers, interrupting SSE2-enabled functions in Chrome, and causing it to crash. This CL restricts the agents from making use of SSE2 instructions allowing SSE2-enabled code to be safely instrumented.

BUG=
R=etienneb@chromium.org, rogerm@chromium.org, siggi@chromium.org

Review URL: https://codereview.appspot.com/97820043