Reorganize ShadowMarkers.
This lifts the ShadowMarker definition out of the Shadow class, placing it on
its own. Due to the increased complexity and number of markers it also adds a
handful of functions for querying, modifying and generating shadow markers,
removing this knowledge from the Shadow implementation itself. Finally,
extensive unittests are added.

BUG=
R=sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/128330043